Implemented stats, working on output formatting issues

+++ cogs/ladder_bot_cog.py
@@ -4,11 +4,17 @@ import discord
 from discord.ext import commands
 from discord import app_commands, Embed
 from datetime import datetime, timedelta
+from table2ascii import table2ascii
 
 class LadderBot_cog(commands.Cog):
     def __init__(self, bot):
         self.bot = bot
         self.load_data()
+
+    @commands.Cog.listener()
+    async def on_ready(self):
+        for guild in self.bot.guilds:
+            await self.update_ladder(guild)
 
     def load_data(self):
         # Read files and initialize variables
@@ -27,29 +33,37 @@ class LadderBot_cog(commands.Cog):
             self.locked_players = data.split('\n')
             self.locked_players.pop(-1)
 
+        with open('./data/ladder/stats.txt','r+') as file:
+            data = file.read()
+            self.stats = data.split('\n')
+            self.stats.pop(-1)
+
     async def update_ladder(self, guild):
         channel_id = 1176131566741762078  # Replace this with your channel ID
         channel = guild.get_channel(channel_id)
 
         if channel:
-            ladder = ''
+            ladder_table = "```\n"
             active_challenges = 'No active challenges'
 
-            # Ladder
-            for person in self.leaderboard:
-                swords = ''
-                rank = 1 + self.leaderboard.index(person)
+            # Check if there are any entries in the leaderboard
+            if self.leaderboard:
+                ladder_table += "{:<5} {:<22} \n".format("Rank", "Player")
+                for person in self.leaderboard:
+                    swords = ''
+                    rank = 1 + self.leaderboard.index(person)
+                    # Check if the person is in the activeChallenges list
+                    for element in self.activeChallenges:
+                        if person in element:
+                            swords = "⚔️"  # Unicode character for crossed swords emoji
+                            
+                    try:
+                        username = guild.get_member(int(person)).display_name
+                        ladder_table += "{:<5} {:<22} \n".format(rank, username + swords)
+                    except:
+                        await channel.send(f'Error while trying to get the username of the user: {person}')
 
-                # Check if the person is in the activeChallenges list
-                for element in self.activeChallenges:
-                    if person in element:
-                        swords = ':crossed_swords:'
-
-                try:
-                    username = guild.get_member(int(person)).display_name
-                    ladder += f'{rank}. {username}{swords}\n'
-                except:
-                    await response.send_message(f'Error while trying to get the username of the user: {person}')
+            ladder_table += "```\n"
 
             # Active Challenges
             if len(self.activeChallenges) > 0:
@@ -64,20 +78,17 @@ class LadderBot_cog(commands.Cog):
                         secondPlayer = guild.get_member(int(secondPlayer)).display_name
                         active_challenges += "{:<22} {:<22} {:<15}\n".format(firstPlayer, secondPlayer, date)
                     except:
-                        await response.send_message(f'Error while trying to get the username of one of these users: {firstPlayer}/{secondPlayer}')
+                        await channel.send(f'Error while trying to get the username of one of these users: {firstPlayer}/{secondPlayer}')
 
                 active_challenges += "```\n"
 
             # Construct the embed
             response = Embed(title='Current Ladder and Active Challenges', color=0x0ccff)
-            response.add_field(name='Current Ladder:', value=ladder, inline=False)
+            response.add_field(name='Current Ladder:', value=ladder_table, inline=False)
             response.add_field(name='Active Challenges:', value=active_challenges, inline=False)
 
             # Find and delete the previous embed
-            async for message in channel.history(limit=10):  # Adjust limit as needed
-                print(message.author)
-                print(guild.me)
-                print(message.embeds)
+            async for message in channel.history(limit=50):  # Adjust limit as needed
                 if message.author == guild.me and message.embeds:
                     for embed in message.embeds:
                         if embed.title == 'Current Ladder and Active Challenges':
@@ -88,6 +99,31 @@ class LadderBot_cog(commands.Cog):
             await channel.send(embed=response)
         else:
             print("Error: Channel not found.")
+
+    def update_stats(self, player: str, win: bool):
+        playerInStats = False
+        for stat in self.stats:
+            if player in stat:
+                playerInStats = True
+                player, wins, losses, streak = stat.split(' - ')
+                if win:
+                    wins = int(wins) + 1
+                    streak = int(streak) + 1
+                else:
+                    losses = int(losses) + 1
+                    streak = 0
+
+                self.stats[self.stats.index(stat)] = (f'{player} - {str(wins)} - {str(losses)} - {str(streak)}')
+
+        if not playerInStats:
+            if win:
+                self.stats.append(f'{player} - 1 - 0 - 1')
+            else:
+                self.stats.append(f'{player} - 0 - 1 - 0')
+
+        with open('./data/ladder/stats.txt', "w+") as file:
+            for stat in self.stats:
+                file.write(stat+'\n')
 
     @app_commands.command(name="ladder", description="Show the current ladder")
     async def ladder(self, interaction):
@@ -173,6 +209,7 @@ class LadderBot_cog(commands.Cog):
                     with open('./data/ladder/activeChallenges.txt', "w+") as file:
                         for activeChallenge in self.activeChallenges:
                             file.write(activeChallenge+'\n')
+
                     response = Embed(title="Challenge scheduled", description=f'Challenge between: \n\n{player.mention} and {interaction.guild.get_member(int(playerAboveId)).mention} \n\nis scheduled to be completed by: {date}')
 
                     await self.update_ladder(interaction.guild)
@@ -209,6 +246,9 @@ class LadderBot_cog(commands.Cog):
                     if playerIndex > otherPlayerIndex:
                         self.leaderboard[winnerIndex] = str(player.id)
                         self.leaderboard[winnerIndex+1] = otherPlayer
+                    
+                    self.update_stats(str(player.id), True)
+                    self.update_stats(otherPlayer, False)
 
                 else:
                     response = Embed(title="Results accepted", description=f'Unlucky {player.mention}, maybe you will win next time', color=0xFF5733)
@@ -216,6 +256,9 @@ class LadderBot_cog(commands.Cog):
                     if playerIndex < otherPlayerIndex:
                         self.leaderboard[winnerIndex] = otherPlayer
                         self.leaderboard[winnerIndex+1] = str(player.id)
+
+                    self.update_stats(otherPlayer, True)
+                    self.update_stats(str(player.id), False)
 
                 self.activeChallenges.remove(challenge)
 
@@ -228,24 +271,34 @@ class LadderBot_cog(commands.Cog):
                     file.truncate(0)
                     for leaderboardEntry in self.leaderboard:
                         file.write(leaderboardEntry+'\n')
+                
+                await self.update_ladder(interaction.guild)
                 break
 
         if noActiveChallenge:
             response = Embed(title="Error", description=f'No active challenge with the player: {player.mention} found', color=0xFF5733)
 
-        await self.update_ladder(interaction.guild)
         await interaction.response.send_message(embed=response)
 
     @app_commands.command(name="join", description="Join the ladder!")
     async def join(self, interaction):
-        player = str(interaction.user.id)
+        alreadyIsInLadder = False
+        player = interaction.user
 
-        self.leaderboard.append(player)
-        with open('./data/ladder/leaderboard.txt','w')as file:
-            for x in self.leaderboard:
-                file.write(x+'\n')
-        
-        response = Embed(title='Player added', description=f'Try not to get wrecked', color=0x0ccff)
+        for leaderboardEntry in self.leaderboard:
+            if str(player.id) in leaderboardEntry:
+                response = Embed(title="Cant join the ladder", description=f'{player.mention}, you are already in the ladder', color=0xff5733)
+                alreadyIsInLadder = True
+
+        if not alreadyIsInLadder:
+            self.leaderboard.append(str(player.id))
+            with open('./data/ladder/leaderboard.txt','w')as file:
+                for leaderboardEntry in self.leaderboard:
+                    file.write(leaderboardEntry +'\n')
+            
+            response = Embed(title='Player added', description=f'Try not to get wrecked', color=0x0ccff)
+            await self.update_ladder(interaction.guild)
+
         await interaction.response.send_message(embed=response)
 
     @app_commands.command(name="add", description="Add a player")
@@ -268,6 +321,8 @@ class LadderBot_cog(commands.Cog):
             with open('./data/ladder/leaderboard.txt', 'w') as file:
                 for entry in self.leaderboard:
                     file.write(entry+'\n')
+            
+            await self.update_ladder(interaction.guild)
 
         await interaction.response.send_message(embed=response)
 
@@ -289,8 +344,10 @@ class LadderBot_cog(commands.Cog):
                         self.activeChallenges.remove(challenge)
                 
                 response = Embed(title="Player removed", description=f'Player {player.mention} removed from the ladder', color=0x0ccff)
+                await self.update_ladder(interaction.guild)
             else:
                 response = Embed(title="Error", description=f'Player {player.mention} not recognized.', color=0xFF5733)
+
         await interaction.response.send_message(embed=response)
 
     @app_commands.command(name="cointoss", description="Toss a coin!")
@@ -372,6 +429,7 @@ class LadderBot_cog(commands.Cog):
                             file.write(f'{leaderboardEntry}\n')
 
                     response=Embed(title='Player Locked', description=f'Player locked untilf further notice', color=0x0ccff)
+                    await self.update_ladder(interaction.guild)
 
         if not foundPlayerInLeaderboard:
             response = Embed(title='Player not found', description=f'Player was not found in the leaderboard')
@@ -382,6 +440,8 @@ class LadderBot_cog(commands.Cog):
 
     @app_commands.command(name="unlock", description="Unlock a player")
     async def unlock(self, interaction, player: discord.User):
+
+        response = Embed(title="Error", description=f"Didnt find {player.mention} in the locked players", color=0x0ccff)
         
         for locked_player in self.locked_players:
             if str(player.id) in locked_player:
@@ -398,9 +458,11 @@ class LadderBot_cog(commands.Cog):
                 with open('./data/ladder/leaderboard.txt', 'w+') as file:
                     for leaderboardEntry in self.leaderboard:
                         file.write(f'{leaderboardEntry}\n')
-                
-                response = Embed(title="Unlocked", description="Player unlocked", color=0x0ccff)
-                await interaction.response.send_message(embed=response)
+
+                response = Embed(title="Unlocked", description=f"{player.mention} unlocked", color=0x0ccff)
+                await self.update_ladder(interaction.guild)
+
+        await interaction.response.send_message(embed=response)
 
     @app_commands.command(name="remove-challenge", description="Removes the challenge which has the player in it")
     async def removechallenge(self, interaction, player: discord.User):
@@ -417,7 +479,6 @@ class LadderBot_cog(commands.Cog):
                         file.write(activeChallenge+'\n')
 
                 response = Embed(title="Challenge removed", description=f'The challenge with the player: {player.mention} was removed', color=0x0ccff)
-
                 await self.update_ladder(interaction.guild)
 
         if noActiveChallenge:
@@ -427,7 +488,11 @@ class LadderBot_cog(commands.Cog):
 
     @app_commands.command(name="update-txt", description="Takes all the txt files and changes the names to ids")
     async def updatetxt(self, interaction):
-        print("lol")
+        print("not implemented")
+
+    @app_commands.command(name="get-oldstats", description="Goes through all the /challenge and /results commands to get all the stats")
+    async def getoldstats(self, interaction):
+        print("not implemented")
 
 async def setup(bot:commands.Bot) -> None:
     await bot.add_cog(LadderBot_cog(bot))
